Move Maven artifact parsing to MavenComponentParser
